Force modded items to have a crosshair.
This has to be done because we have no way of differentiating between
regular items and tools from mods. Therefore, the crosshair would be
missing with modded tools and weapons.

Currently, this mod's focus is to be compatible with Vanilla. Having
proper individual mod compatibility would require major architectural
changes. If those happen, it will be for a future version.