Begin rework: Replace logic with tags and ducks
Rip out all of the "instanceof" logic and replace it with a three-way
approach. First level is item/block/entitytype tags, second level is a
duck interface, and third level is the old API system.

This should get rid of the potential evaluation performance issues, so
also remove the "simple" mode. All other options are retained for now.