Convert Crosshair to a holder of two interaction modes
It now stores a "primary interaction" (left click) and a "secondary
interaction" (right click). The mapping of interaction types to actual
crosshairs is now done in CrosshairComponent, so the API is free of it.

Hopefully, all of the intermittent regressions have been fixed by this.

For now, the mapping does not differentiate between e.g. "use item" and
"use item on block", but this can now easily be added at any point
without touching the API. Also, crosshair computation does not have to
consider visual displays anymore.